Move PeerList and its builder into their own module
